app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
jinja2==3.1.6
python-multipart==0.0.20
requests==2.32.3
chromadb==0.5.23
faster-whisper==1.1.1
ffmpeg-python==0.2.0
pydantic==2.11.3
pydantic-settings==2.8.1
python-slugify==8.0.4
